azure_functions/requirements.txt
azure-functions
pymssql==2.2.7
requests
azure-storage-blob